src/backend/requirements.txt
Django==4.2.10
mssql-django==1.4.2
pyodbc==5.3.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
